+++ Source/Context/IntegratorAI.Context.Domain/ToolParameter.cs
@@ -2,8 +2,8 @@
 
 public class ToolParameter
 {
-    public long Id { get; set; }
-    public long ToolId { get; set; }
+    public int Id { get; set; }
+    public int ToolId { get; set; }
     public string Name { get; set; } = null!;
     public string Type { get; set; } = null!;
     public string Description { get; set; } = null!;
